Done Migrations, Updated the customer models with customer id, also updated menu models to display menu id, also tweaked coffeeshop

+++ BFGCoffeeShop.Models/MenuModels/MenuCreate.cs
@@ -10,6 +10,6 @@ namespace BFGCoffeeShop.Models.MenuModels
     {
         public string ItemName { get; set; }
 
-        public double ItemPrice { get; set; }
+        public decimal ItemPrice { get; set; }
     }
 }
